+++ pages/user/balance/add.js
@@ -1,43 +1,19 @@
-import axios from "axios";
 import { useEffect, useRef, useState } from "react"
-import { Button, Card, Form } from "react-bootstrap";
-import Header from '../../components/Header'
+import { Button } from "react-bootstrap";
+import useScript from '../../hooks/useScript'
 
-const PUBLIC_KEY = "TEST-2724ac84-5ce2-45d6-8de4-df9d8378ac77";
+const PUBLIC_KEY = "TEST-001debb2-d8d5-40a4-953f-8ca65aaa0fa0";
 
-export default function Index() {
+export default function Index(props) {
   //const { id } = useParams(); // id de producto
   const id = 1122;
-  const [preferenceId, setPreferenceId] = useState(null)
-  const [price, setPrice] = useState('');
 
+  const [preferenceId, setPreferenceId] = useState(null)
+  const formRef = useRef(null)
   const initialized = useRef(false)
+
   const { loaded } = useScript("https://sdk.mercadopago.com/js/v2")
 
-  function useScript(url) {
-    const [loaded, setLoaded] = useState(false)
-    useEffect(() => {
-      const existingScript = document.querySelector(`[src="${url}"]`)
-  
-      if (existingScript) {
-        setLoaded(true)
-      } else {
-        const script = document.createElement("script")
-        script.src = url
-        script.async = true
-        script.onload = () => {
-          setLoaded(true)
-        }
-  
-        document.body.appendChild(script)
-      }
-    }, [url])
-  
-    return {
-      loaded
-    }
-  }
-  
   useEffect(() => {
     if (initialized.current) {
       console.log('initialized.current')
@@ -56,7 +32,7 @@ export default function Index() {
 
     initialized.current = true
 
-    const mp = new window.MercadoPago(PUBLIC_KEY, {
+    const mp = new window.MercadoPago('TEST-2724ac84-5ce2-45d6-8de4-df9d8378ac77', {
       locale: "pt-BR",
     });
 
@@ -73,36 +49,8 @@ export default function Index() {
   }, [loaded, preferenceId])
 
 
-
-  const fetchPayment = async (event) => {
-    event.preventDefault();
-    event.stopPropagation();
-    try {
-      const { data } = await axios.post(
-        "http://localhost:3000/api/user/balance/add",
-        { 
-          productId: id, 
-          description: "Adicionar saldo na carteira SG",
-          price,
-          quantity: 1, },
-        {
-          headers: {
-            "Content-Type": "application/json",
-            "Access-Control-Allow-Origin": "*",
-          },
-        }
-      );
-      setPreferenceId(await data.id);
-      console.log('data.id', data.id)
-      console.log('try')
-      return data;
-    } catch (error) {
-      console.log('catch', error)
-    }
-  };
-
-/*     const fetchId = async () => {
-      console.log('fetch')
+  useEffect(() => {
+    const fetchId = async () => {
       try {
         const post = await fetch("http://localhost:3000/api/user/balance/add", {
           method: "POST",
@@ -113,8 +61,8 @@ export default function Index() {
           },
           body: JSON.stringify({
             productId: id,
-            description: "Adicionar saldo na carteira SG",
-            price,
+            description: "agustin",
+            price: 1,
             quantity: 1,
           }),
         });
@@ -126,32 +74,15 @@ export default function Index() {
       } catch (error) {
         console.log('error', error);
       }
-    } */
+    }
 
+    fetchId()
 
-
- 
+  }, []);
 
   return (
     <div>
-      <Header />
-      <Card>
-      <Form.Group>
-          <Form.Label>Valor : </Form.Label>
-          <Form.Control
-            onChange={({ target: { value } }) => setPrice(value)}
-            name="valor"
-            type="number"
-            placeholder="10,00"
-          />
-        </Form.Group>
-      </Card>
-      <div className='d-grid gap-2 mt-2'>
-        <Button onClick={fetchPayment}>
-           Gerar Botão de pagamento
-        </Button>
-        </div>
-     <center><div className="button-checkout" id="button-checkout"></div></center>
+      <div className="button-checkout" id="button-checkout"></div>
     </div>
   )
   ;
